♻️ D43531 context: decompose command dispatcher
