+++ .agentplane/tasks/202605282222-D43531/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 D43531 task: Context command dispatcher decomposition [202605282222-D43531]
