+++ .agentplane/tasks/202605282222-D43531/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T22:23:10.540Z
+
+## Task
+
+- Task: `202605282222-D43531`
+- Title: Context command dispatcher decomposition
+- Status: DOING
+- Branch: `task/202605282222-D43531/context-command-dispatcher-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282222-D43531/README.md`
+
+## Verification
+
+- State: ok
+- Note: Context command dispatcher split into thin entrypoint, runner, group usage handlers, and interactive init runner. Verified with focused context CLI tests, typecheck, arch deps, lint, format, and hotspot threshold check (runtime warnings 40 -> 39).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T22:23:10.540Z
+- Branch: task/202605282222-D43531/context-command-dispatcher-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
